chore: refactor native android common
