+++ android/src/main/java/com/swmansion/enriched/textinput/utils/EnrichedSpannableStringBuilder.kt
@@ -1,6 +1,7 @@
 package com.swmansion.enriched.textinput.utils
 
 import android.text.SpannableStringBuilder
+import com.swmansion.enriched.common.EnrichedConstants
 
 // Removes zero-width spaces from the given range in the SpannableStringBuilder without affecting spans
 fun SpannableStringBuilder.removeZWS(
